Rename remaining Optimize component files to new naming conventions.
