+++ assets/js/modules/optimize/settings/SettingsMain.js
@@ -25,8 +25,8 @@ import { useEffect } from '@wordpress/element';
  * Internal dependencies
  */
 import Data from 'googlesitekit-data';
-import SettingsEdit from './settings-edit';
-import SettingsView from './settings-view';
+import SettingsEdit from './SettingsEdit';
+import SettingsView from './SettingsView';
 import { STORE_NAME } from '../datastore/constants';
 const { useSelect, useDispatch } = Data;
