+++ assets/js/modules/optimize/settings/SettingsMain.test.js
@@ -21,7 +21,7 @@
  */
 import { render, fireEvent, wait } from '../../../../../tests/js/test-utils';
 import { STORE_NAME } from '../datastore/constants';
-import SettingsMain from './settings-main';
+import SettingsMain from './SettingsMain';
 
 describe( 'SettingsMain', () => {
 	const initialSettings = {
