+++ assets/js/modules/optimize/settings/SettingsView.js
