+++ assets/js/modules/optimize/setup/index.js
@@ -16,5 +16,5 @@
  * limitations under the License.
  */
 
-export { default as SetupMain } from './setup-main';
-export { default as SetupForm } from './setup-form';
+export { default as SetupMain } from './SetupMain';
+export { default as SetupForm } from './SetupForm';
